+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Features
 
  - Add Input: TTN Integration: Data Storage (TTN v3), replaces deprecated TTN v2 Input (deprecated 12/2021)
+ - Add Input: HTU21D Variant using the CircuitPython library
 
 
 ## 8.12.8 (2021-10-27)